# KuryeNet/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KuryeNet API
  version: 0.0.1
servers:
  - url: http://localhost:8585
paths:
...
  /api/vehicles/getByCourierId:
    get:
      summary: getByCouriers_CourierId
      parameters:
        - in: query
          name: courierId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/vehicles/getByCourierIdentityNumber:
    get:
      summary: getByCouriers_CourierIdentityNumber
      parameters:
        - in: query
          name: courierIdentityNumber
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/vehicles/getByCourierNameAndCourierSurname:
    get:
      summary: getByCouriers_CourierNameAndCouriers_CourierSurname
      parameters:
        - in: query
          name: courierName
          schema: { type: string }
        - in: query
          name: courierSurname
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/vehicles/getByVehicleBrand:
    get:
      summary: getByVehicleBrand
      parameters:
        - in: query
          name: vehicleBrand
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/vehicles/getByVehicleId:
    get:
      summary: getByVehicleId
      parameters:
        - in: query
          name: vehicleId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/vehicles/getByVehicleModel:
    get:
      summary: getByVehicleModel
      parameters:
        - in: query
          name: vehicleModel
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/vehicles/getByVehiclePlate:
    get:
      summary: getByVehiclePlate
      parameters:
        - in: query
          name: vehiclePlate
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/vehicles/getByVehicleType:
    get:
      summary: getByVehicleType
      parameters:
        - in: query
          name: vehicleType
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/vehicles/getByVehicleYear:
    get:
      summary: getByVehicleYear
      parameters:
        - in: query
          name: vehicleYear
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/vehicles/getallByPage:
    get:
      summary: getAll
      parameters:
        - in: query
          name: pageNo
          schema: { type: string }
        - in: query
          name: pageSize
          schema: { type: string }
      responses:
        '200': { description: OK }
...